feat: only install requirements from required python extensions to speed up use/build processing (#451)
* fix: remove BUILD.gn for python extensions

* feat: only install requirements if extension exists in manifest

* fix: remove llama index preload

* feat: reset in repo python extension versions to 0.1.0

* feat: add necessary py extensions in manifest

+++ agents/scripts/install_deps_and_build.sh
@@ -60,7 +60,11 @@ install_python_requirements() {
   if [[ -d "ten_packages/extension" ]]; then
     for extension in ten_packages/extension/*; do
       if [[ -f "$extension/requirements.txt" ]]; then
-        pip install -r $extension/requirements.txt
+        local extension_name=$(basename $extension)
+        local exists_in_manifest=$(cat manifest.json | jq --arg ext_name "${extension_name}" 'any(.dependencies[].name == $ext_name ; .)' )
+        if [[ "$exists_in_manifest" == "true" ]]; then
+          pip install -r $extension/requirements.txt
+        fi
       fi
     done
   fi
@@ -73,10 +77,6 @@ install_python_requirements() {
       fi
     done
   fi
-
-  # pre-import llama-index as it cloud download additional resources during the first import
-  echo "pre-import python modules..."
-  python3.10 -c "import llama_index.core;"
 }
 
 build_go_app() {
